langgraph_oracle_project/langgraph_oracle_project/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
oracledb>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
